eslint collectiblesgallery.vue
